Limit max trace depth to 255 using 8-bit unsigned integer

+++ coretrace/simulation_runner/optix_runner/OptixCSP/src/core/ray_utils.h
@@ -14,7 +14,7 @@ namespace OptixCSP
     /// as long as num_rays and max_depth stay the same.
     struct CompactionScratch
     {
-        uint32_t *d_count = nullptr;   // per-ray output record count
+        uint8_t  *d_count = nullptr;   // per-ray output record count (bounded by max_depth ≤ 255)
         uint64_t *d_offsets = nullptr; // exclusive prefix-sum of d_count (pass 1); reused as global ray IDs (pass 2)
         uint8_t  *d_has_hit = nullptr; // 1 if ray contributes records, else 0
         uint32_t *d_n_hit = nullptr;   // scalar: total hit rays
